normalize zones & mask point

+++ frigate/config.py
@@ -479,12 +479,12 @@ class ZoneConfig(BaseModel):
 
         if isinstance(coordinates, list):
             self._contour = np.array(
-                [[int(p.split(",")[0]), int(p.split(",")[1])] for p in coordinates]
+                [[float(p.split(",")[0]), float(p.split(",")[1])] for p in coordinates]
             )
         elif isinstance(coordinates, str):
             points = coordinates.split(",")
             self._contour = np.array(
-                [[int(points[i]), int(points[i + 1])] for i in range(0, len(points), 2)]
+                [[float(points[i]), float(points[i + 1])] for i in range(0, len(points), 2)]
             )
         else:
             self._contour = np.array([])
